added logout component and restructured the app
